build: rename --enable-teamdctl configure option to --enable-team
--enable* options are used to specify optional features and in this
case the feature is support for team devices, so --enable-team sounds
more appropriate, instead of --enable-teamdctl which is the name of
the library required to control teamd.

At the moment the presence of the library is the only requirement
needed to build the plugin, but in next commits new requirements will
be added, making the current option name misleading.