libnm-core: improve matching of static routes metric
When we compare a generated connection with a persistent one at
startup, the metric field of static routes is also considered.
However, if the field is set to -1 (default) on the persistent
connection, the comparison will always fail because the generated
connection contains the actual value read from kernel.

The effect of this is that NM fails to re-assume the previous
connection after a restart when static routes are present.

To fix the issue, let's match default metric values with any other
value when the NM_SETTING_COMPARE_FLAG_INFERRABLE flag is passed.

https://bugzilla.redhat.com/show_bug.cgi?id=1302532